# Add timer to thinking indicator in agent mode