--- a/src/ui/agent-view.ts
+++ b/src/ui/agent-view.ts
@@ -1071,13 +1071,17 @@ These files are included in the context below. When the user asks you to write d
 					// Use streaming API with tool support
 					let modelMessageContainer: HTMLElement | null = null;
 					let accumulatedMarkdown = '';
-					
-					// Remove thinking indicator before streaming starts
-					thinkingMessage.remove();
-					this.stopTimer();
+					let thinkingRemoved = false;
 					
 					const streamResponse = modelApi.generateStreamingResponse(request, (chunk: string) => {
 						accumulatedMarkdown += chunk;
+						
+						// Remove thinking indicator when first chunk arrives
+						if (!thinkingRemoved) {
+							thinkingMessage.remove();
+							this.stopTimer();
+							thinkingRemoved = true;
+						}
 						
 						// Create or update the model message container
 						if (!modelMessageContainer) {
@@ -1104,6 +1108,13 @@ These files are included in the context below. When the user asks you to write d
 						
 						// Check if the model requested tool calls
 						if (response.toolCalls && response.toolCalls.length > 0) {
+							// Remove thinking indicator if it hasn't been removed yet
+							if (!thinkingRemoved) {
+								thinkingMessage.remove();
+								this.stopTimer();
+								thinkingRemoved = true;
+							}
+							
 							// Save user message to history first
 							if (this.plugin.settings.chatHistory) {
 								await this.plugin.sessionHistory.addEntryToSession(this.currentSession, userEntry);
@@ -1159,6 +1170,13 @@ These files are included in the context below. When the user asks you to write d
 								console.warn('Model returned empty response');
 								new Notice('Model returned an empty response. This might happen with thinking models. Try rephrasing your question.');
 								
+								// Remove thinking indicator if it hasn't been removed yet
+								if (!thinkingRemoved) {
+									thinkingMessage.remove();
+									this.stopTimer();
+									thinkingRemoved = true;
+								}
+								
 								// Still save the user message to history
 								if (this.plugin.settings.chatHistory) {
 									await this.plugin.sessionHistory.addEntryToSession(this.currentSession, userEntry);
@@ -1168,6 +1186,11 @@ These files are included in the context below. When the user asks you to write d
 					} catch (error) {
 						this.currentStreamingResponse = null;
 						this.plugin.settings.sendContext = originalSendContext;
+						// Remove thinking indicator if it hasn't been removed yet
+						if (!thinkingRemoved) {
+							thinkingMessage.remove();
+							this.stopTimer();
+						}
 						throw error;
 					}
 				} else {