--- a/styles.css
+++ b/styles.css
@@ -286,9 +286,10 @@ If your plugin does not need CSS, delete this file.
 
 .gemini-scribe-timer {
 	color: var(--text-muted);
-	font-size: 14px;
+	font-size: 0.9em;
+	font-family: var(--font-monospace);
 	margin-top: 5px;
-	min-width: 4em;
+	min-width: 3.5em;
 	text-align: center;
 }
 
@@ -2108,6 +2109,7 @@ If your plugin does not need CSS, delete this file.
 	font-style: italic;
 }
 
+/* Thinking animation dots */
 .gemini-agent-thinking-dot {
 	display: inline-block;
 	width: 4px;
@@ -2117,20 +2119,21 @@ If your plugin does not need CSS, delete this file.
 	animation: thinking-bounce 1.4s infinite ease-in-out both;
 }
 
-.gemini-agent-timer {
-	color: var(--text-muted);
-	font-size: 0.9em;
-	font-family: var(--font-monospace);
-	min-width: 3.5em;
-	text-align: right;
-}
-
 .gemini-agent-thinking-dot-1 {
 	animation-delay: -0.32s;
 }
 
 .gemini-agent-thinking-dot-2 {
 	animation-delay: -0.16s;
+}
+
+/* Timer display */
+.gemini-agent-timer {
+	color: var(--text-muted);
+	font-size: 0.9em;
+	font-family: var(--font-monospace);
+	min-width: 3.5em;
+	text-align: right;
 }
 
 @keyframes thinking-bounce {