## Summary
Fixes #118 - Adds a timer display to the thinking indicator in agent mode, improving the user experience by showing elapsed time while waiting for model responses.

## Problem
The agent mode had two issues:
1. No timer showing elapsed time during API calls (just animated dots)
2. The thinking indicator was being removed too quickly before users could see it

## Solution
1. Added a timer that shows elapsed seconds (e.g., "0.0s", "1.5s", "10.2s") next to the thinking animation
2. Fixed the timing of when the thinking indicator is removed - now it stays visible until actual content arrives

## Changes
### Timer Implementation
- Added timer properties (`timerInterval` and `startTime`) to `AgentView` class
- Modified thinking indicator HTML structure to include timer display
- Implemented `startTimer()` and `stopTimer()` methods
- Timer updates every 100ms for smooth display

### Visibility Fix
- Moved thinking indicator removal from "before streaming starts" to "when first content chunk arrives"
- Added proper cleanup for all scenarios (tool calls, empty responses, errors)
- Ensures users can actually see the thinking indicator working

## Visual Changes
Before: `Thinking...` (animated dots only, disappeared too quickly)
After: `Thinking... 2.3s` (animated dots + timer, stays visible until content arrives)

## Testing
- Built and tested the plugin locally
- Verified timer displays and updates correctly during API calls
- Confirmed thinking indicator stays visible until content arrives
- Tested cleanup on view close and error scenarios
- All existing tests pass

## Impact
This improvement provides better user feedback during AI processing, making the agent mode feel more responsive and informative, especially during longer operations or when using slower models.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Introduces a shared ChatTimer and integrates an accessible elapsed-time display into Agent thinking and Gemini views, deferring indicator removal until content arrives and updating styles/cleanup.
> 
> - **UI**
>   - **Agent (`src/ui/agent-view.ts`)**:
>     - Thinking indicator now includes a timer (`aria-live`, `aria-label`) and new containers: `gemini-agent-thinking-container`, `gemini-agent-thinking-text-container`, `gemini-agent-timer`.
>     - Defers removal of the thinking indicator until the first streamed chunk or tool handling; ensures `ChatTimer.stop()` on normal completion, empty responses, errors, and non-streaming paths; destroys timer on view close.
>   - **Gemini (`src/ui/gemini-view.ts`)**:
>     - Replaces inline timer logic with `ChatTimer`; starts on send, stops and hides after completion; adds accessible timer display; destroys timer on close.
> - **Utils**
>   - **New `src/utils/timer-utils.ts`**: `ChatTimer` class with `start`, `stop`, `isRunning`, `getElapsedTime`, `destroy` for 100ms updates.
> - **Styles (`styles.css`)**
>   - Add/adjust styles for agent and scribe timers (monospace, 0.9em, min-width), thinking containers, and timing dot animations.
> 
> <sup>Written by [Cursor Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit 1d070babef3f6b6255a066a1ec15ca8b77134dc5. This will update automatically on new commits. Configure [here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->